fix: comm (campioni lag), contatori connessione, hardening act.info (scan/look/where)
comm.cpp / comm.hpp
- Corretto uso di aTimeCheck[] e NumTimeCheck: prima NumTimeCheck era inizializzato
  a sizeof(aTimeCheck) (dimensione in byte dell’array, non lunghezza) e l’indice usava % sizeof(aTimeCheck), fuori semantica rispetto a PULSE_MOBILE.
- Ora NumTimeCheck è un indice ad anello che parte da 0 e gli slot sono
  aTimeCheck[NumTimeCheck % PULSE_MOBILE], con commenti aggiornati in .hpp.

act.info.hpp / act.info.cpp
- HowManyConnection: ritorno std::uint64_t, contatore static unsigned; incremento solo per ToAdd > 0 e decremento sicuro per ToAdd < 0 (no sotto‑zero).
- signals.cpp: mudlog allineato con %llu e cast per il nuovo tipo di ritorno.
- singular: bool, const struct obj_data*, guardia su nullptr.
- argument_split_2: controlli nullptr su argument e buffer di uscita.
- get_object_in_equip_vis, find_ex_description: const‑correctness e guardie.
- do_where_object: parametro recurse da int a bool (firma in .hpp e call site).
- show_obj_to_char / show_char_to_char (e affini): estrazione helper in namespace anonimo (es. etichette aura, equipaggiamento visibile, capitalizzazione), meno duplicazione e percorsi più chiari.
- do_scan: cammino lineare unificato (linear_walk + visitor), scan_emit_*,
  controlli su ch / stanza; skip loopback verso la stanza di partenza dove serve.
- can_see_linear / get_char_linear: stessa logica di cammino dei visitor usati
  dallo scan, con callback per blocchi/errori.
- list_groups: controlli ch/desc, messaggi di log più coerenti su follower nulli.
- CheckCharAffected: rimosso prototipo orfano; aggiunto blocco commento storico
  (Nebbie) con il void commentato e motivazione della rimozione.

Altri comandi in act.info.cpp (do_look, do_examine, do_status, do_achievements, do_score, ecc.): revisione diffusa con guardie, semplificazioni e refactor locali
(oltre 2k righe nette anche ignorando solo‑whitespace).

+++ src/signals.cpp
@@ -54,8 +54,8 @@ void PrintStatus() {
 }
 void PrintStatus(int level) {
 	int i=0;
-	mudlog(LOG_SYSERR, "Connections from start: %d",
-		   HowManyConnection(0));
+	mudlog(LOG_SYSERR, "Connections from start: %llu",
+		   static_cast<unsigned long long>(HowManyConnection(0)));
 	if(level==1) {
 		mudlog(LOG_SYSERR,"CurrentTrack %s at %d",currentfile,currentline);
 		mudlog(LOG_SYSERR, "Mud status: '%s'",gszMudStatus);
